Set daphne ping timeout in miner to 20min
Daphne misuses `autobahn` and autopings are causing
disconnects after ping message is dropped by client.
Other pings/pongs still work, but after lost ping timeout
connection is closed by server. This interrupts validator
synthetic batch. We set it to 20min so it will not occure during
batch.

This is hack and we either need to fix it in daphne or swith
daphne to other asgi server that handles pings correctly.
Independently we will investigate validator code (why
ping message is lost), but this should be handled independently.