Switch to use main hotkey instead of split distribution (#606)
Initial implementation of the new dancing mechanism assumed that a miner would return a split distribution for all hotkeys under its coldkey. To simplify this, the miner now only needs to return its main hotkey. The main hotkey receives a configured share (MAIN_HOTKEY_SHARE) of the total score for the coldkey. Dancing is applied to the total score of a given coldkey, not just to the main hotkey.