fix: temporarily avoid archive subtensor when backfilling
Allowance backfilling has trouble catching up and hardly crosses the threshold
of blocks usable for organic jobs.

This makes it so that the backfilling starts in a block range that doesn't
require archive data.

This should be reverted when a proper fix is implemented for the original issue,
as this may cause a different set of allowance blocks being processed across
different validators.

Impacts: validator