fix: aget_config blocking the receipts transfer
Somehow we ended up getting stuck on the retrieval of Constance config,
but only in the "keep up" part of the receipts transfer. It would only
run the initial catch-up procedure and then do nothing, sitting and
waiting for aget_config to return.

I have not dug deep into why exactly it was blocked. Instead, I moved
switched the aget_config helper to use `thread_sensitive=False` - it
should not be a problem for retrieving Constance configs. We should
address the async issues globally anyway.

Impacts: validator