Fix receipt transfer (#767)
* fix: sync sleep in async receipts transfer

This never seemed to manifest itself as an actual problem, but still was
 fishy.

 Impacts: validator

* fix: receipt transfer getting stuck on async stuff

Somehow the receipts transfer procedure keeps getting stuck on async
calls like aget_config and the async miner source.

I have not dug deep into why exactly it was blocked. Instead, I
changed up the retrieval to not use the database which seems to have
shuffled things around and we're not getting stuck any more. For a
proper fix we must address the async issues globally.

Impacts: validator