fix(validator): Fix redis connection corruption
Moving AsyncResult to a different thread was corrupting the redis
connection. This fix moves the construction of AsyncResult in the thread
that waits for it.

It *should* fix the following issues:
- https://sentry.reef.pl/organizations/reef-technologies/issues/740279/
- https://sentry.reef.pl/organizations/reef-technologies/issues/740104/
- https://sentry.reef.pl/organizations/reef-technologies/issues/740608/

Here's why I think it is happenning:

The redis backend of celery is not thread safe. Celery uses thread local
storage to store a backend for each thread:
https://github.com/celery/celery/blob/v5.3.6/celery/app/base.py#L1278-L1302

We are constructing the `AsyncResult` in one thread, which grabs the
backend for that specific thread from the app:
https://github.com/celery/celery/blob/v5.3.6/celery/result.py#L96

But we are sending that `AsyncResult` object (more accurately, the bound
method) to another thread (`sync_to_async(thread_sensitive=False)`) to
call the `get()` method which in turn tries to listen on the redis
connection, corrupting it in the process.

Impacts: validator