fix(validator): separate metrics of prometheus-proxy from app (#829)
The multiprocs db files are being treated as corrupted when read from
app's /metrics endpoint, but not when read from prometheus-proxy itself.
This results in the following error when serving the metrics from app:
https://sentry.reef.pl/organizations/reef-technologies/issues/742050/

This is not a proper fix, but a workaround. This makes prometheus
additionally scrape the /metrics endpoint of prometheus-proxy and
moves the multiproc dir of prometheus-proxy out of app's central
./prometheus-metrics.

Impacts: validator